putting all the errors with not found element to be handled by NotFound.js

+++ apiexpress/src/controllers/authorControllers.js
@@ -1,3 +1,5 @@
+import NotFound from '../errors/NotFound.js';
+//import fetchElement from '../middlewares/fetchElement.js';
 import { author } from '../models/Author.js';
 // import mongoose from 'mongoose';
 
@@ -15,8 +17,9 @@ class AuthorController {
         try {
             const id = req.params.id;
             const fetchedAuthor = await author.findById(id);
+            console.log(author);
             if (!fetchedAuthor) {
-                res.status(400).json({message : 'author id not found'});
+                next(new NotFound('author id not found'));
             } else {
                 res.status(200).json(fetchedAuthor);
             }
@@ -37,10 +40,16 @@ class AuthorController {
     static async updateAuthorById(req, res, next) {
         try {
             const id = req.params.id;
-            await author.findByIdAndUpdate(id, req.body);
-            res
-                .status(200)
-                .json({ message: `Author Updated: ${author.findById(id)}` });
+            const fetchedAuthor = await author.findById(id);
+            console.log(author);
+            if (!fetchedAuthor) {
+                next(new NotFound('author id not found'));
+            } else {
+                await author.findByIdAndUpdate(id, req.body);
+                res
+                    .status(200)
+                    .json({ message: `Author Updated: ${Object.values(await author.findById(id)).map(author => author.name).join('')}` });
+            }
         } catch (error) {
             next(error);
         }
@@ -48,12 +57,20 @@ class AuthorController {
     static async deleteAuthorById(req, res, next) {
         try {
             const id = req.params.id;
-            await author.findByIdAndDelete(id);
-            res
-                .status(200)
-                .json({
-                    message: `author successfully deleted: ${author.findById(id)}`,
-                });
+            const fetchedAuthor = await author.findById(id);
+            console.log(author);
+            if (!fetchedAuthor) {
+                next(new NotFound('author id not found'));
+            } else {
+                const authorName = Object.values(await author.findById(id)).map(author => author.name).join('');
+                await author.findByIdAndDelete(id);
+                res
+                    .status(200)
+                    .json({
+                        message: `author successfully deleted: ${authorName}`,
+                    });
+            }
+            
         } catch (error) {
             next(error);
         }
